permissions is functional and all reference checks and data validation is in place

+++ src/api/permissions/permissions.js
@@ -1,9 +1,8 @@
-import jsonPatch from 'jsonpatch';
 import Boom from '@hapi/boom';
 import dal from './dal';
+import roles from '../roles/roles';
 import helper from '../../helper';
 import ueEvents from '../../events/ueEvents';
-import access from '../accounts/access';
 
 export default {
 	async writePermission(data) {
@@ -22,8 +21,27 @@ export default {
 	},
 
 	async deletePermission(authGroupId, product, id) {
+		const permission = await dal.getPermission(authGroupId, product, id);
+		if(!permission) throw Boom.notFound(`id requested was ${id}`);
+		const coded = `${id} ${permission.coded}`;
+		const rolesCleared = await roles.clearPermission(authGroupId, product, coded);
 		const result = await dal.deletePermission(authGroupId, product, id);
-		ueEvents.emit(authGroupId, 'ue.permission.destroy', result);
+		const output = { rolesImpacted: rolesCleared, permission: result };
+		ueEvents.emit(authGroupId, 'ue.permission.destroy', output);
+		return output;
+	},
+
+	async deletePermissionsByProduct(authGroupId, product) {
+		const result = await dal.deletePermissionsByProduct(authGroupId, product);
+		ueEvents.emit(authGroupId, 'ue.permission.destroy', { authGroup: authGroupId, product, result });
 		return result;
+	},
+
+	async checkForProductReference(authGroup, product) {
+		const result = await dal.checkForProductReference(authGroup, product);
+		return {
+			totalReferences: result.length,
+			permissionIds: result
+		};
 	}
 };
